requirements.txt
# 功能测试用例生成器依赖包
# 已验证可正常安装和运行

# Web框架
Flask>=2.3.0

# 数据处理 - 确保numpy和pandas版本兼容
numpy>=1.21.0,<2.0.0
pandas>=2.0.0

# HTTP请求
requests>=2.31.0

# Excel处理
openpyxl>=3.1.0

# MySQL数据库
PyMySQL>=1.1.0

# 从项目根目录 .env 加载环境变量（可选，便于本地配置 MYSQL_*）
python-dotenv>=1.0.0

# LLM 返回的畸形 JSON 修复（严格生成器回退解析）
json-repair>=0.30.0

# 可选：AI服务支持
# openai>=1.0.0
# anthropic>=0.3.0
